+++ src/App.jsx
@@ -9,6 +9,7 @@ import SelectControlDocs from './docs/select-control.mdx';
 import RangeControlDocs from './docs/range-control.mdx';
 import TextControlDocs from './docs/text-control.mdx';
 import DateTimePickerDocs from './docs/date-time-picker.mdx';
+import CheckboxControlDocs from './docs/checkbox-control.mdx';
 import ColorPaletteDocs from './docs/color-palette.mdx';
 
 /**
@@ -28,6 +29,7 @@ export default function App() {
 				<Route path="/docs/range-control" element={ <RangeControlDocs /> } />
 				<Route path="/docs/text-control" element={ <TextControlDocs /> } />
 				<Route path="/docs/date-time-picker" element={ <DateTimePickerDocs /> } />
+				<Route path="/docs/checkbox-control" element={ <CheckboxControlDocs /> } />
 				<Route path="/docs/color-palette" element={ <ColorPaletteDocs /> } />
 			</Routes>
 		</DocLayout>