+++ src/components/index.js
@@ -1,6 +1,7 @@
 export { default as Alert } from './Alert';
 export { default as Button } from './Button';
 export { default as Card } from './Card';
+export { default as CheckboxControl } from './CheckboxControl';
 export { default as CodeBlock } from './CodeBlock';
 export { default as CodeTabs } from './CodeTabs';
 export { default as PropsTable } from './PropsTable';